added config.json, detect installed tools in Config, finished mingw install in Build

+++ haxe/Build.hx
@@ -1,5 +1,7 @@
 class Build {
 
+	static var CFG = @:privateAccess Config.CFG;
+
 	var cygwinPath : String;
 
 	function new() {
@@ -27,8 +29,27 @@ class Build {
 		command("bash",["-c",cmd+" "+args.join(" ")]);
 	}
 
-	function cygCopyFile( file : String, out : String ) {
-		try sys.io.File.copy(cygwinPath + "/" + file, "mingw/" + out + "/" + file) catch( e : Dynamic ) log("*** MISSING " + file+" in your Cygwin install ***");
+	var copiedFiles = new Map();
+
+	function cygCopyFile( file : String ) {
+
+		copiedFiles.set(file, true);
+		try sys.io.File.copy(cygwinPath + "/" + file, "mingw/" + file) catch( e : Int ) log("*** MISSING " + file+" in your Cygwin install ***");
+
+		if( !StringTools.endsWith(file.toLowerCase(),".exe") )
+			return;
+
+		// look for dll dependencies
+		var o = new sys.io.Process("dumpbin.exe",["/IMPORTS","mingw/"+file]);
+		var lines = o.stdout.readAll().toString().split("\n");
+		o.exitCode();
+		var r = ~/^[A-Za-z0-9_-]+\.dll$/;
+		for( f in lines ) {
+			var f = StringTools.trim(f);
+			if( !r.match(f) || copiedFiles.exists(f) ) continue;
+			if( !sys.FileSystem.exists(cygwinPath+"/bin/"+f) ) continue;
+			cygCopyFile("bin/"+f);
+		}
 	}
 
 	function detectCygwin() {
@@ -37,7 +58,33 @@ class Build {
 			log("Cygwin not found");
 			Sys.exit(1);
 		}
-		cygwinPath = StringTools.trim(p.stdout.readAll().toString()).substr(0,-11);
+		cygwinPath = StringTools.trim(p.stdout.readAll().toString()).substr(0,-15);
+	}
+
+	function cygCopyDir( dir : String ) {
+		var baseDir = "";
+		for( f in dir.split("/") ) {
+			if( baseDir == "" )
+				baseDir = f;
+			else
+				baseDir += "/"+f;
+			makeDir("mingw/"+baseDir);
+		}
+		cygCopyRec(dir);
+	}
+
+	function cygCopyRec( dir : String ) {
+		var cygDir = cygwinPath + "/" + dir;
+		var files = try sys.FileSystem.readDirectory(cygDir) catch( e : Dynamic ) {
+			log("*** MISSING "+dir+" in your Cygwin install ***");
+			return;
+		}
+		for( f in files )
+			if( sys.FileSystem.isDirectory(cygDir+"/"+f) ) {
+				makeDir("mingw/"+dir+"/"+f);
+				cygCopyRec(dir+"/"+f);
+			} else
+				cygCopyFile(dir+"/"+f);
 	}
 
 	function build() {
@@ -45,26 +92,51 @@ class Build {
 		// build minimal mingw distrib to use if the user doesn't have cygwin installed
 
 		detectCygwin();
+		Sys.println("Preparing mingw distrib...");
 		makeDir("mingw/bin");
-		for( f in Config.REQUIRED_TOOLS )
-			cygCopyFile(f + ".exe", "bin");
-
-		return;
+		for( f in CFG.cygwinTools )
+			cygCopyFile("bin/"+ f + ".exe");
+		cygCopyDir("usr/i686-w64-mingw32");
+		cygCopyDir("lib/gcc/i686-w64-mingw32");
+		makeDir("mingw/tmp");
 
 		// build opam repo with packages necessary for haxe
 
+		Sys.println("Preparing opam...");
+
 		var opam = "opam32.tar.xz";
-		var ocaml = "4.03.0+mingw32";
+		var ocaml = CFG.ocamlVersion + "+mingw32";
 
-		deleteFile(opam);
-		command("wget",["https://github.com/fdopen/opam-repository-mingw/releases/download/0.0.0.1/"+opam]);
-		command("tar",["-xf",opam,"--strip-components","1"]);
-		deleteFile(opam);
-		deleteFile("install.sh");
+		if( !sys.FileSystem.exists("bin") ) {
+			// install opam
+			deleteFile(opam);
+			command("wget",["https://github.com/fdopen/opam-repository-mingw/releases/download/0.0.0.1/"+opam]);
+			command("tar",["-xf",opam,"--strip-components","1"]);
+			deleteFile(opam);
+			deleteFile("install.sh");
+		}
 
-		var opamPath = Sys.getCwd().split("\\").join("/")+".opam";
-		cygCommand("bin/opam",["init","--yes","--root",opamPath,"default","https://github.com/fdopen/opam-repository-mingw.git","--comp",ocaml,"--switch",ocaml]);
-		cygCommand("bin/opam",["install","--yes","--root",opamPath,"sedlex","camlp4","merlin";"extlib";"xml-light";"camlzip"]);
+		// copy necessary runtime files so they are added to PATH in Config
+		for( lib in CFG.mingwLibs ) {
+			var out = "bin/"+lib+".dll";
+			if( !sys.FileSystem.exists(out) )
+				sys.io.File.copy("mingw/usr/i686-w64-mingw32/sys-root/mingw/bin/"+lib+".dll",out);
+		}
+
+		var cwd = Sys.getCwd().split("\\").join("/");
+		var opamRoot = cwd+".opam";
+
+		// temporarily modify the env
+		Sys.putEnv("PATH", cwd+"bin;" + Sys.getEnv("PATH"));
+		Sys.putEnv("OPAMROOT", opamRoot);
+
+		if( !sys.FileSystem.exists(opamRoot) )
+			cygCommand("opam",["init","--yes","default","https://github.com/fdopen/opam-repository-mingw.git","--comp",ocaml,"--switch",ocaml]);
+
+		cygCommand("opam",["switch",ocaml]);
+		cygCommand("opam",["install","--yes"].concat(CFG.opamLibs));
+
+		Sys.println("DONE");
 	}
 
 	static function main() {
